+++ CHANGELOG.rst
@@ -12,6 +12,13 @@ Changelog
 * add plot functions to ``sscalc`` and ``torsions``
 * PR #198
 
+v0.5.0 (2022-05-24)
+------------------------------------------------------------
+
+* Add residue tolerance matrices: EDSS50
+* Update/improve parameters to residue tolerance options
+* PR #183
+
 v0.4.10 (2022-05-23)
 ------------------------------------------------------------